backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-ai
cohere
python-dotenv
